requirements.txt
fastapi
uvicorn
python-dotenv
camel-ai
pillow
requests-oauthlib
pandas